==> 06 Batsmen Stats.Rmd <==
---
title: "06 Batsmen Stats"
author: "Arlton Gilbert"
date: "23/05/2021"
output: html_document
---

```{r setup, include=FALSE}
#FILENUM: 6
knitr::opts_chunk$set(echo = TRUE)
#knitr::opts_knit$set(root.dir = dirname(rstudioapi::getActiveDocumentContext()$path))
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
library(tidyverse)
library("xlsx")
library("readxl")
library(zoo) #rollaply
library(openxlsx) #normal excel wasn't writing 
rm(list=ls())
#Average Run Time ~ 20.74min

```

```{r clearprevious}
#ensure old files are not left there
do.call(file.remove, list(list.files("Indiv Stats Ext/Bat Stats Ext", full.names = TRUE)))
do.call(file.remove, list(list.files("Indiv Stats Ext/Group Bat Stats Ext", full.names = TRUE)))

```

```{r loadings}
#To get the batsmen names
load(file='02 Player Roles.RData') #roles 
load(file='05 Player Splitter.RData') #batnames
start.time <- Sys.time()
sink("Indiv Stats Ext/Bat Log.txt")

```

```{r names}

mom = 5 #number of matches to use in momentum
mu = 0.8

#Player Roles
#colnames(batext) #to copy the names
stat_heads = c("Player","MP","MB","NotOut","R","B","4s","6s","SR",
"Extras","Stadium","Date","Time","Team","Opponent","InningsNum","BatNum","TeamScore","OppScore",
"Toss","Decision","MatchNum","Result","Captain","WicketKeeper","Role","Cleaned Role","Country","Matches Played","Matches Batted","Tot Not Outs","Tot Runs","Tot Balls","Tot 4s","Tot 6s","Unedited Average","Average","Strike Rate","Career Bat Score","Weighted Career Bat Score","Running Matches Played","Running Matches Batted","Running Not Outs","Running Runs","Running Balls","Running Unedited Average","Running Average","Running Strike Rate","Running Bat Score","Weighted Running Bat Score","Batting Strength","Weighted Batting Strength")

#Overall Stats
ncols = length(stat_heads)
all_bat_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
batsman_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
bat_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
bowl_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
bowler_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)

#Indian Players
local_all_bat_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
local_batsman_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
local_bat_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
local_bowl_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
local_bowler_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)

#Foreign Players
foreign_all_bat_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
foreign_batsman_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
foreign_bat_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
foreign_bowl_allrounder_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)
foreign_bowler_bat_scores=setNames(data.frame(matrix(ncol=ncols,nrow=0)),stat_heads)

for (name in batnames$Player){
  batinfo = read_excel(paste0("Indiv Stats/Bat Stats/",name," Bat.xlsx"),sheet = "Bat")
  role = profilered[tolower(profilered$Player) == tolower(name), "Position"]
  country = profilered[tolower(profilered$Player) == tolower(name), "Country"]
  
  print(paste0('processing ',name))
  
  if (is.na(role) || role == "--"){
    print (paste0(name,' not found in the player roles'))
  }
  cleaned_role = str_replace(role,'WK-','')
  
  if (is.na(country)){
    print (paste0(name,' not found in the countries'))
  } 
  
  batext = batinfo %>% add_column(Role = role, `Cleaned Role` = cleaned_role, Country = country,
    `Matches Played` = cumsum(batinfo$MP), 
    `Matches Batted` = cumsum(batinfo$MB),
    `Tot Not Outs` = cumsum(ifelse(is.na(batinfo$NotOut),0,batinfo$NotOut)),
    `Tot Runs` = cumsum(ifelse(is.na(batinfo$R),0,batinfo$R)),
    `Tot Balls` = cumsum(ifelse(is.na(batinfo$B),0,batinfo$B)),
    `Tot 4s` = cumsum(ifelse(is.na(batinfo$`4s`),0,batinfo$`4s`)),
    `Tot 6s` = cumsum(ifelse(is.na(batinfo$`6s`),0,batinfo$`6s`)),
    `Unedited Average` = `Tot Runs`/(`Matches Batted` - `Tot Not Outs`),
    Average = ifelse((`Matches Batted` - `Tot Not Outs`)>0,`Tot Runs`/(`Matches Batted` - `Tot Not Outs`),`Tot Runs`),
    `Strike Rate` = ifelse(`Tot Balls`>0,(`Tot Runs`/`Tot Balls`)*100,0),
    `Career Bat Score` = Average * `Strike Rate`,
    `Weighted Career Bat Score` = sqrt(`Matches Batted`/`Matches Played`)*`Career Bat Score`,
    `Running Matches Played` = rollapply(ifelse(is.na(batinfo$`MP`),0,batinfo$`MP`),mom,sum,partial = TRUE,align='right'),
    `Running Matches Batted` = rollapply(ifelse(is.na(batinfo$`MB`),0,batinfo$`MB`),mom,sum,partial = TRUE,align='right'),
    `Running Not Outs` = rollapply(ifelse(is.na(batinfo$NotOut),0,batinfo$NotOut),mom,sum,partial = TRUE,align='right'),
    `Running Runs` = rollapply(ifelse(is.na(batinfo$R),0,as.numeric(batinfo$R)),mom,sum,partial = TRUE,align='right'),
    `Running Balls` = rollapply(ifelse(is.na(batinfo$B),0,as.numeric(batinfo$B)),mom,sum,partial = TRUE,align='right'),
    `Running Unedited Average` = `Running Runs`/(`Running Matches Batted` - `Running Not Outs`),
    `Running Average` = ifelse((`Running Matches Batted` - `Running Not Outs`)>0,`Running Runs`/(`Running Matches Batted` - `Running Not Outs`),`Running Runs`),
    `Running Strike Rate` = ifelse(`Running Balls`>0,(`Running Runs`/`Running Balls`)*100,0),
    `Running Bat Score` = `Running Average` * `Running Strike Rate`,
    `Weighted Running Bat Score` = sqrt(`Running Matches Batted`/`Running Matches Played`) * `Running Bat Score`,
    `Batting Strength` = mu * `Career Bat Score` + (1- mu)*`Running Bat Score`,
    `Weighted Batting Strength` = mu * `Weighted Career Bat Score` + (1- mu)*`Weighted Running Bat Score`)
  #weighted means weighted by matches batted/matches played
  #Running means last x (currently 5) matches
  #Overall takes mu (currently 0.8) of career and rest in running
  #Unedited Average is infinite if they've never been out and the edited Average is just the total runs

  assign(paste0(str_replace_all(name," ","_"),'_bat_ext'),batext)  
  
  all_bat_bat_scores= rbind(all_bat_bat_scores,batext)
  if(batext$`Cleaned Role`[1] == 'Batsman'){
    batsman_bat_scores= rbind(batsman_bat_scores,batext)
  }
  else if(batext$`Cleaned Role`[1] == 'Batting Allrounder'){
    bat_allrounder_bat_scores= rbind(bat_allrounder_bat_scores,batext)
  }
  else if(batext$`Cleaned Role`[1] == 'Bowling Allrounder'){
    bowl_allrounder_bat_scores= rbind(bowl_allrounder_bat_scores,batext)
  }
  else if(batext$`Cleaned Role`[1] == 'Bowler'){
    bowler_bat_scores= rbind(bowler_bat_scores,batext)
  }

  if(batext$Country[1] == "India"){
    local_all_bat_bat_scores= rbind(local_all_bat_bat_scores,batext)
    if(batext$`Cleaned Role`[1] == 'Batsman'){
      local_batsman_bat_scores= rbind(local_batsman_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Batting Allrounder'){
      local_bat_allrounder_bat_scores= rbind(local_bat_allrounder_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Bowling Allrounder'){
      local_bowl_allrounder_bat_scores= rbind(local_bowl_allrounder_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Bowler'){
      local_bowler_bat_scores= rbind(local_bowler_bat_scores,batext)
    }
  }
  
  else{
    foreign_all_bat_bat_scores= rbind(foreign_all_bat_bat_scores,batext)
    if(batext$`Cleaned Role`[1] == 'Batsman'){
      foreign_batsman_bat_scores= rbind(foreign_batsman_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Batting Allrounder'){
      foreign_bat_allrounder_bat_scores= rbind(foreign_bat_allrounder_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Bowling Allrounder'){
      foreign_bowl_allrounder_bat_scores= rbind(foreign_bowl_allrounder_bat_scores,batext)
    }
    else if(batext$`Cleaned Role`[1] == 'Bowler'){
      foreign_bowler_bat_scores= rbind(foreign_bowler_bat_scores,batext)
    }
  }
  
  print(paste0(name," done"))
}

#View(batext)
#View(all_bat_bat_scores)
#View(batsman_bat_scores)
#View(bat_allrounder_bat_scores)
#View(bowl_allrounder_bat_scores)
#View(bowler_bat_scores)
#View(local_bowler_bat_scores)
#View(foreign_bowler_bat_scores)
#View(local_batsman_bat_scores)
#View(foreign_batsman_bat_scores)

sink()

```

```{r bestbat}
#best batsmen with at least 5 games
best_bat = all_bat_bat_scores %>% filter(`Matches Played`>=mom) %>% arrange(desc(`Batting Strength`)) %>% distinct(Player,.keep_all = TRUE) %>% head(20) 
View(best_bat)

weighted_best_bat = all_bat_bat_scores %>% filter(`Matches Played`>=mom) %>% arrange(desc(`Weighted Batting Strength`)) %>% distinct(Player,.keep_all = TRUE) %>% head(20)
View(weighted_best_bat)

```


```{r batsave}
openxlsx::write.xlsx(all_bat_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/All Bat Scores.xlsx",sheetName ="All Bat Scores", append = FALSE)
openxlsx::write.xlsx(batsman_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Batsman Bat Scores.xlsx",sheetName ="Batsman Bat Scores", append = FALSE)
openxlsx::write.xlsx(bat_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Batting Allrounder Bat Scores.xlsx",sheetName ="Bat All Bat Scores", append = FALSE)
openxlsx::write.xlsx(bowl_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Bowling Allrounder Bat Scores.xlsx",sheetName ="Bowl All Bat Scores", append = FALSE)
openxlsx::write.xlsx(bowler_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Bowler Bat Scores.xlsx",sheetName ="Bowler Bat Scores", append = FALSE)

#Indian Players
openxlsx::write.xlsx(local_all_bat_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Local All Bat Scores.xlsx",sheetName ="All Bat Scores", append = FALSE)
openxlsx::write.xlsx(local_batsman_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Local Batsman Bat Scores.xlsx",sheetName ="Batsman Bat Scores", append = FALSE)
openxlsx::write.xlsx(local_bat_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Local Batting Allrounder Bat Scores.xlsx",sheetName ="Bat All Bat Scores", append = FALSE)
openxlsx::write.xlsx(local_bowl_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Local Bowling Allrounder Bat Scores.xlsx",sheetName ="Bowl All Bat Scores", append = FALSE)
openxlsx::write.xlsx(local_bowler_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Local Bowler Bat Scores.xlsx",sheetName ="Bowler Bat Scores", append = FALSE)

#Foreign Players
openxlsx::write.xlsx(foreign_all_bat_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign All Bat Scores.xlsx",sheetName ="All Bat Scores", append = FALSE)
openxlsx::write.xlsx(foreign_batsman_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign Batsman Bat Scores.xlsx",sheetName ="Batsman Bat Scores", append = FALSE)
openxlsx::write.xlsx(foreign_bat_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign Batting Allrounder Bat Scores.xlsx",sheetName ="Bat All Bat Scores", append = FALSE)
openxlsx::write.xlsx(foreign_bowl_allrounder_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign Bowling Allrounder Bat Scores.xlsx",sheetName ="Bowl All Bat Scores", append = FALSE)
openxlsx::write.xlsx(foreign_bowler_bat_scores, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign Bowler Bat Scores.xlsx",sheetName ="Bowler Bat Scores", append = FALSE)

print("Writing done")

```

```{r filter}
# only take those who have played enough matches
all_bat_bat_clear = all_bat_bat_scores %>% filter(`Running Matches Played` == mom)
batsman_bat_clear = batsman_bat_scores %>% filter(`Running Matches Played` == mom)
bat_allrounder_bat_clear = bat_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
bowl_allrounder_bat_clear = bowl_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
bowler_bat_clear = bowler_bat_scores %>% filter(`Running Matches Played` == mom)

local_all_bat_bat_clear = local_all_bat_bat_scores %>% filter(`Running Matches Played` == mom)
local_batsman_bat_clear = local_batsman_bat_scores %>% filter(`Running Matches Played` == mom)
local_bat_allrounder_bat_clear = local_bat_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
local_bowl_allrounder_bat_clear = local_bowl_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
local_bowler_bat_clear = local_bowler_bat_scores %>% filter(`Running Matches Played` == mom)

foreign_all_bat_bat_clear = foreign_all_bat_bat_scores %>% filter(`Running Matches Played` == mom)
foreign_batsman_bat_clear = foreign_batsman_bat_scores %>% filter(`Running Matches Played` == mom)
foreign_bat_allrounder_bat_clear = foreign_bat_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
foreign_bowl_allrounder_bat_clear = foreign_bowl_allrounder_bat_scores %>% filter(`Running Matches Played` == mom)
foreign_bowler_bat_clear = foreign_bowler_bat_scores %>% filter(`Running Matches Played` == mom)

```

```{r matchstats}
bat_col_names = c("Real Match Num","MatchNum","Avg All Bat Strength","Avg Batsman Bat Strength","Avg Bat Allrounder Bat Strength","Avg Bowl Allrounder Bat Strength","Avg Bowler Bat Strength","Avg All Bat Weighted Strength","Avg Batsman Bat Weighted Strength","Avg Bat Allrounder Bat Weighted Strength","Avg Bowl Allrounder Bat Weighted Strength","Avg Bowler Bat Weighted Strength")
col_tots = length(bat_col_names)

overall_bat_summary = setNames(data.frame(matrix(ncol=col_tots,nrow=tot_matches+1)),bat_col_names)
local_overall_bat_summary = setNames(data.frame(matrix(ncol=col_tots,nrow=tot_matches+1)),bat_col_names)
foreign_overall_bat_summary = setNames(data.frame(matrix(ncol=col_tots,nrow=tot_matches+1)),bat_col_names)

score = function(bat_clear){
  return (mean((bat_clear %>% filter(MatchNum <= matches) %>% group_by(Player) %>% summarise(across(everything(), last)))$`Batting Strength`)) #test different sections of this
}

weight_score = function(bat_clear){
  return (mean((bat_clear %>% filter(MatchNum <= matches) %>% group_by(Player) %>% summarise(across(everything(), last)))$`Weighted Batting Strength`)) #test different sections of this
}

#Match 0
overall_bat_summary[,1] = c(0:tot_matches)
overall_bat_summary[,2] = c(1:(tot_matches+1))

local_overall_bat_summary[,1] = c(0:tot_matches)
local_overall_bat_summary[,2] = c(1:(tot_matches+1))

foreign_overall_bat_summary[,1] = c(0:tot_matches)
foreign_overall_bat_summary[,2] = c(1:(tot_matches+1))

for (matches in c(1:tot_matches)){
  # take only those matches up to our current match, then only take the last match in that of each player and then the average Batting_Strength of all these
  #MatchNumber is the number for merging since we want as at the end of the previous match
  overall_bat_summary[matches+1,3] = score(all_bat_bat_clear)
  overall_bat_summary[matches+1,4] = score(batsman_bat_clear)
  overall_bat_summary[matches+1,5] = score(bat_allrounder_bat_clear)
  overall_bat_summary[matches+1,6] = score(bowl_allrounder_bat_clear)
  overall_bat_summary[matches+1,7] = score(bowler_bat_clear)
  
  overall_bat_summary[matches+1,8] = weight_score(all_bat_bat_clear)
  overall_bat_summary[matches+1,9] = weight_score(batsman_bat_clear)
  overall_bat_summary[matches+1,10] = weight_score(bat_allrounder_bat_clear)
  overall_bat_summary[matches+1,11] = weight_score(bowl_allrounder_bat_clear)
  overall_bat_summary[matches+1,12] = weight_score(bowler_bat_clear)
  
  #Indian Players
  local_overall_bat_summary[matches+1,3] = score(local_all_bat_bat_clear)
  local_overall_bat_summary[matches+1,4] = score(local_batsman_bat_clear)
  local_overall_bat_summary[matches+1,5] = score(local_bat_allrounder_bat_clear)
  local_overall_bat_summary[matches+1,6] = score(local_bowl_allrounder_bat_clear)
  local_overall_bat_summary[matches+1,7] = score(local_bowler_bat_clear)
  
  local_overall_bat_summary[matches+1,8] = weight_score(local_all_bat_bat_clear)
  local_overall_bat_summary[matches+1,9] = weight_score(local_batsman_bat_clear)
  local_overall_bat_summary[matches+1,10] = weight_score(local_bat_allrounder_bat_clear)
  local_overall_bat_summary[matches+1,11] = weight_score(local_bowl_allrounder_bat_clear)
  local_overall_bat_summary[matches+1,12] = weight_score(local_bowler_bat_clear)
  
  #Indian Players
  foreign_overall_bat_summary[matches+1,3] = score(foreign_all_bat_bat_clear)
  foreign_overall_bat_summary[matches+1,4] = score(foreign_batsman_bat_clear)
  foreign_overall_bat_summary[matches+1,5] = score(foreign_bat_allrounder_bat_clear)
  foreign_overall_bat_summary[matches+1,6] = score(foreign_bowl_allrounder_bat_clear)
  foreign_overall_bat_summary[matches+1,7] = score(foreign_bowler_bat_clear)
  
  foreign_overall_bat_summary[matches+1,8] = weight_score(foreign_all_bat_bat_clear)
  foreign_overall_bat_summary[matches+1,9] = weight_score(foreign_batsman_bat_clear)
  foreign_overall_bat_summary[matches+1,10] = weight_score(foreign_bat_allrounder_bat_clear)
  foreign_overall_bat_summary[matches+1,11] = weight_score(foreign_bowl_allrounder_bat_clear)
  foreign_overall_bat_summary[matches+1,12] = weight_score(foreign_bowler_bat_clear)
  
  print(paste0(matches," done"))
}

View(overall_bat_summary)
View(local_overall_bat_summary)
View(foreign_overall_bat_summary)

```

```{r batsave2}
openxlsx::write.xlsx(overall_bat_summary, file = "Indiv Stats Ext/Group Bat Stats Ext/Overall Bat Summary.xlsx",sheetName ="Overall Bat Summary", append = FALSE)
openxlsx::write.xlsx(local_overall_bat_summary, file = "Indiv Stats Ext/Group Bat Stats Ext/Local Overall Bat Summary.xlsx",sheetName ="Overall Bat Summary", append = FALSE)
openxlsx::write.xlsx(foreign_overall_bat_summary, file = "Indiv Stats Ext/Group Bat Stats Ext/Foreign Overall Bat Summary.xlsx",sheetName ="Overall Bat Summary", append = FALSE)

```

```{r BatExt}
add_scores = function(bat_summary,var1,var2){
    file_overall = file2 %>% merge(bat_summary[,c("MatchNum",var1,var2)],by = "MatchNum") 
    file_overall_sort = file_overall[order(file_overall$`Matches Played`),]
    file_overall_sort$`Net Bat Strength` = file_overall_sort$`Running Matches Played`/mom * file_overall_sort$`Batting Strength` + (1 - file_overall_sort$`Running Matches Played`/mom) * file_overall_sort[[var1]]
    file_overall_sort$`Net Weighted Bat Strength` = file_overall_sort$`Running Matches Played`/mom * file_overall_sort$`Weighted Batting Strength` + (1 - file_overall_sort$`Running Matches Played`/mom) * file_overall_sort[[var2]]
    return(file_overall_sort)
}

#change to have a sepearate file for each player like before
for (name in batnames$Player){
  file = get(paste0(str_replace_all(name," ","_"),'_bat_ext'))
  rows=dim(file)[1]
  #View(file)
  split_point = which(stat_heads == 'Country')[[1]]#29
  filelead = rbind(file[,c(1:split_point)],0)
  fileoffset = rbind(0,file[,c((split_point+1):ncols)])
  file2 = cbind(filelead,fileoffset)
  file2[rows+1,2] = file2[rows,2]
  #View(file2)
  
  cleaned_role = str_replace(profilered[tolower(profilered$Player) == tolower(name), "Position"],'WK-','')
  country = profilered[tolower(profilered$Player) == tolower(name),"Country"]

  if(country == "India"){
    bat_summary = local_overall_bat_summary
  }
  else{
    bat_summary = foreign_overall_bat_summary
  }
  
  if(cleaned_role == "Batsman"){
    ext_batting_info = add_scores(bat_summary,"Avg Batsman Bat Strength","Avg Batsman Bat Weighted Strength")
    }
  else if(cleaned_role == "Batting Allrounder"){
    ext_batting_info = add_scores(bat_summary,"Avg Bat Allrounder Bat Strength","Avg Bat Allrounder Bat Weighted Strength")
  }
  else if(cleaned_role == "Bowling Allrounder"){
    ext_batting_info = add_scores(bat_summary,"Avg Bowl Allrounder Bat Strength","Avg Bowl Allrounder Bat Weighted Strength")
  }
  else if(cleaned_role == "Bowler"){
    ext_batting_info = add_scores(bat_summary,"Avg Bowler Bat Strength","Avg Bowler Bat Weighted Strength")
  }
  
  new_file_name = paste0(str_replace_all(name," ","_"),'_bat_ext') 
  assign(paste0(str_replace_all(name," ","_"),'_bat_ext'),ext_batting_info)
  openxlsx::write.xlsx(ext_batting_info, file = paste0("Indiv Stats Ext/Bat Stats Ext/",new_file_name,".xlsx"),sheetName ="Bat Stats Ext", append = FALSE)
}

View(ext_batting_info)

```

```{r savestate}
end.time <- Sys.time()
time.taken <- end.time - start.time
time.taken

#Savestate so we don't have to rerun computationally intensive parts later
save.image(file='06 Batsmen Stats.RData')

```

```{r loadstate}
#load(file='06 Batsmen Stats.RData')

```

```{r runabove}
```
